Move support for async functions as operation handlers into non-underscored packages.
