+++ Sources/SmokeOperationsHTTP1/SmokeHTTP1HandlerSelector+withInputWithOutput.swift
@@ -12,10 +12,10 @@
 // permissions and limitations under the License.
 //
 // SmokeHTTP1HandlerSelector+withInputWithOutput.swift
-// _SmokeOperationsHTTP1Concurrency
+// SmokeOperationsHTTP1
 //
 
-#if compiler(>=5.5)
+#if compiler(>=5.5) && canImport(_Concurrency)
 
 import Foundation
 import SmokeOperations
