+++ Sources/_SmokeOperationsConcurrency/Export.swift
@@ -1,0 +1,19 @@
+// Copyright 2018-2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
+//  AdditionalHeadersOperationHTTPOutput.swift
+//  _SmokeOperationsConcurrency
+//
+
+// TODO: https://github.com/amzn/smoke-framework/issues/85
+@_exported import SmokeOperations
